Correct expand_k type from int to float for gla config

+++ fla/models/gla/configuration_gla.py
@@ -13,7 +13,7 @@ class GLAConfig(PretrainedConfig):
     def __init__(
         self,
         hidden_size: int = 2048,
-        expand_k: int = 0.5,
+        expand_k: float = 0.5,
         expand_v: int = 1,
         hidden_ratio: Optional[int] = 4,
         intermediate_size: Optional[int] = None,
